Home: props user & state isLogged
The idea is to set a timer if the user is not logged for playing, and if the user is logged, displaying his dashboard, on the Home page

So the first step was to fetch the user data via props and including it in the Home component for being able to check if user is logged or not. The setter associated is generated with useState.

Another idea is to present the Home page in a Grid layout for a easier styles' handling.
Just a test for seeing the rendering of a thing like that, will be improve after