Home: inserting a clock with effect
Create a Clock component, and a Timer component that renders the current time and the current time front of Home page without being logged.

The Clock updates each second for fetching accurate time. formatting the output if the num is under 10.
have Needed to create a useInterval hook for using setInterval and clearInterval (https://overreacted.io/fr/making-setinterval-declarative-with-react-hooks/)

This concerns only the code's logic not the displaying