Home: creating a DigitalClock component
Time is displaying as a clock would do.
A red Border is only here to see where the component is, it will be deleted.
Visitor can be stressed if we need, like to hurry up for login, it can be deactivated. It is still raw with the style, but the clock woks and updates.