Home: isNotLoggedGrid
The logic is finished, it could be update if any ideas are submitted, the style is not defined yet, only the positions of the box are set, trying the most to do responsive disposition. Any suggestions are welcomed. But the logic should be this one submitted. Adding a Fade effect for adding the comp is not logged. Remove the component if we are logged a dashboard should appear