Home: Dashboard
items can be added, updated and deleted by column the one after the others according to a grid system.
console.log are still there for debugging and checking that events bubble well where they have to.
The idea is to have a Customizable Dashboard that appears when User logs in. Do not know yet what to display but it can be some recent activities, or friends' activities, or best scores, or stats or whatever. This is not finished.

remains to determine which datas to display and fetch them. remains to keep the settings of the user (json file, bdd?)
remains to add the possibility to delete a column of datas; to add an autocomplete field for letting the user which theme/item it would like to be displayed.
remains to improve the forms, their positions and how they display datas, maybe if it is not too ambitious doing a drag and drop feat for the items. remains to improve the collections that will keep datas.
Borders are drawn only for marking their boxes, it is more easier to see which settings to update for adjusting styles to the viewer (tried to make it the most responsive as possible